pyproject.toml
[project]
name = "research-studio"
version = "0.1.0"
description = "LangGraph-powered deep research agent with web search, iterative gap detection, and source verification."
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-tavily>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
# Better HTML extraction (recommended)
extract = [
    "trafilatura>=1.6.0",
    "beautifulsoup4>=4.12.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
